move loading spinner into RegisterForm
